added negative number support

+++ source/messages.h
@@ -18,5 +18,5 @@
 #define MESSAGE_POS_Y 21
 
 mqd_t init_queue (char *mq_name, int open_flags);
-void send_message(mqd_t mq, uint16_t command, uint16_t value);
-int get_message(mqd_t mq, uint16_t *command, uint16_t *value);
+void send_message(mqd_t mq, uint16_t command, int value);
+int get_message(mqd_t mq, uint16_t *command, int *value);
